\name{copa-package}
\alias{copa-package}
\docType{package}
\title{
copa - A package to compute 'Cancer Outlier Profile Analysis'
}
\description{
This package is used to compute copa scores, p-values based on
permutation, and plots of paired genes.
}
\details{
\tabular{ll}{
Package: \tab copa\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2006-01-26\cr
License: \tab LLGPL \cr
}
There are two main functions; \code{copa}, which is used to compute the COPA
score for a set of microarrays, and \code{permCopa}, which is used to
calculate permutation based p-values and estimate false discovery rate (FDR).
}
\author{
James W. MacDonald

Maintainer: James W. MacDonald <jmacdon@med.umich.edu>
}
\references{
Tomlins, SA, et.al. Recurrent fusion of TMPRSS2 and ETS transcription
factor genes in prostate cancer.
Science. 2005 Oct 28;310(5748):644-8. 
}
\keyword{ package }
