\name{perm.mat}
\alias{perm.mat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Produce a Matrix of Permuted Classlabels}
\description{
  This function makes a matrix of permuted classlabels. This is not
  intended to be called directly by end users.
}
\usage{
perm.mat(B, ids)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{B}{ The number of permutations}
  \item{ids}{A vector of classlabels}
}

\value{
 A matrix of permuted classlabels.
}

\author{James W. MacDonald}

\keyword{ internal}
