\name{pSum}
\alias{pSum}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Compute all pairwise sums}
\description{
  A function that computes all pairwise sums for a vector of
  numbers. This is an internal function and is not intended for use by end-users.
}
\usage{
pSum(a)
}

\arguments{
  \item{a}{ A numeric vector}

\value{
 \item{out}{A square matrix (of dimension length(a) X length(a))
   containing all pairwise sums.
}

\author{ James W. MacDonald}

\keyword{ internal }% at least one, from doc/KEYWORDS

